issue #235 - added zoom support for video.
