+++ cam2/src/main/java/com/commonsware/cwac/cam2/CameraActivity.java
@@ -52,14 +52,6 @@ public class CameraActivity extends AbstractCameraActivity
     "cwac_cam2_save_preview";
 
   /**
-   * Extra name for whether the camera should allow zoom and
-   * how. Value should be a ZoomStyle (NONE, PINCH, SEEKBAR).
-   * Default is NONE.
-   */
-  public static final String EXTRA_ZOOM_STYLE=
-    "cwac_cam2_zoom_style";
-
-  /**
    * Extra name for how much heap space we should try to use
    * to load the picture for the confirmation screen. Should
    * be a `float` greater than 0.0f and less than 1.0f.
@@ -289,18 +281,6 @@ public class CameraActivity extends AbstractCameraActivity
 
     public IntentBuilder debugSavePreviewFrame() {
       result.putExtra(EXTRA_DEBUG_SAVE_PREVIEW_FRAME, true);
-
-      return(this);
-    }
-
-    /**
-     * Call to configure the ZoomStyle to be used. Default
-     * is NONE.
-     *
-     * @return the builder, for further configuration
-     */
-    public IntentBuilder zoomStyle(ZoomStyle zoomStyle) {
-      result.putExtra(EXTRA_ZOOM_STYLE, zoomStyle);
 
       return(this);
     }
