+++ cam2/src/main/java/com/commonsware/cwac/cam2/VideoRecorderActivity.java
@@ -72,6 +72,7 @@ public class VideoRecorderActivity extends AbstractCameraActivity {
     return(CameraFragment.newVideoInstance(getOutputUri(),
         getIntent().getBooleanExtra(EXTRA_UPDATE_MEDIA_STORE, false),
         getIntent().getIntExtra(MediaStore.EXTRA_VIDEO_QUALITY, 1),
+        (ZoomStyle)getIntent().getSerializableExtra(EXTRA_ZOOM_STYLE),
         getIntent().getIntExtra(MediaStore.EXTRA_SIZE_LIMIT, 0),
         getIntent().getIntExtra(MediaStore.EXTRA_DURATION_LIMIT, 0),
         getIntent().getBooleanExtra(EXTRA_FACING_EXACT_MATCH, false),
